+++ tests/test_cookies.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 
 import pytest
+from urllib.parse import unquote
 
 from blacksheep import (
     Cookie,
@@ -235,9 +236,15 @@ def test_datetime_from_cookie_format_2(expected_result, value):
         ),
         (
             b"NID=204=0K7PurlER1icDcU_vBBCFWff0gPjtSX3saNz-AXBmkjWGi7RWl_XEeV4uAUuHdX0qsAJbaAhl8E-fZTjwMlTyB9Du_bkal2PHdlnz6h0iKsBNjC5ee8JePM-0PW6hCKdyxyORH6Dzhd7kkvJBhZzk6HQz0QeP8vi9h9eDGL0RGs; expires=Mon, 22-Feb-2021 11:13:40 GMT; path=/; domain=.google.com; HttpOnly",
-            "NID=204",
-            "0K7PurlER1icDcU_vBBCFWff0gPjtSX3saNz-AXBmkjWGi7RWl_XEeV4uAUuHdX0qsAJbaAhl8E-fZTjwMlTyB9Du_bkal2PHdlnz6h0iKsBNjC5ee8JePM-0PW6hCKdyxyORH6Dzhd7kkvJBhZzk6HQz0QeP8vi9h9eDGL0RGs",
+            "NID",
+            "204=0K7PurlER1icDcU_vBBCFWff0gPjtSX3saNz-AXBmkjWGi7RWl_XEeV4uAUuHdX0qsAJbaAhl8E-fZTjwMlTyB9Du_bkal2PHdlnz6h0iKsBNjC5ee8JePM-0PW6hCKdyxyORH6Dzhd7kkvJBhZzk6HQz0QeP8vi9h9eDGL0RGs",
             "/",
+        ),
+        (
+            b"session=gAAAAABgVeIWAXQ5iCbIgXThcx9IFORha534yIqw2ZjnqiTKIw7xBcnk-Tc8pvuTpLEuFSv3NRJkr83WBdhc0dpjZrEGBUNCFV8YK17hka43KanCxW5FMhrP00AxvGYyKZ2-vy4CEUIcsN92JAvV763u_ZCZzSpraw==",
+            "session",
+            "gAAAAABgVeIWAXQ5iCbIgXThcx9IFORha534yIqw2ZjnqiTKIw7xBcnk-Tc8pvuTpLEuFSv3NRJkr83WBdhc0dpjZrEGBUNCFV8YK17hka43KanCxW5FMhrP00AxvGYyKZ2-vy4CEUIcsN92JAvV763u_ZCZzSpraw==",
+            None,
         ),
     ],
 )